refactor: breakdown-page
Details:
- Follow the "ducks" pattern for organizing the components' file structure
