+++ workspaces/redhat-resource-optimization/plugins/redhat-resource-optimization/src/components/InfoCardTitle.tsx
