+++ workspaces/redhat-resource-optimization/plugins/redhat-resource-optimization/src/pages/optimizations-breakdown/components/optimization-engine-tab/components/chart-info-card/components/intl-breakdown-chart/components/optimizations-breakdown-chart/components/optimizations-breakdown-chart/i18n/intl.ts
