+++ workspaces/redhat-resource-optimization/plugins/redhat-resource-optimization/src/pages/optimizations-breakdown/components/optimization-engine-tab/components/chart-info-card/components/intl-breakdown-chart/components/optimizations-breakdown-chart/components/optimizations-breakdown-chart/utils/chart-data-format.ts
@@ -25,7 +25,7 @@ import {
   RecommendationType,
   ResourceType,
   UsageType,
-} from '../types/ChartEnums';
+} from '../../../../../../../../../../../models/ChartEnums';
 import { format } from 'date-fns';
 
 export const getRecommendationTerm = (
